[TMP]: Moving SPI and TWI interfaces to new SDK.
Conversion is partly complete. There are files that still need to be
changed to use the new "nrf_drv_twi.h" and "nrf_drv_spi.h" functions.

So far, migrated:
* db.c
* fb.c
* spi.c
* mpu6050.c

Additionally, I changed the configuration file to enable TWI, SPI and
the timer module. I will try reverting to a previous commit and just
enabling the timer module to see if that solves the issue with the BLDC
motor controller.