[FIX]: Messaging now works for all faceboards.
There are 6 buffers, one for each faceboard. They are 128 bytes long, so
messages will now conform to a standard 128 byte length which is yet to
be decided upon. There are about 80 bytes of memory left total out of
16KB, so this solution is not sustainable.

To get this to work, I had to remove a bunch of strings in commands.c
that generally corresponded to unused or least used functionality.

Will try to convert to SDKv7 in the mean time, but at least this is a
temporary solution that works.